# skill-openapi3-spec.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "3호선 도착 정보 확인",
        "description": "OpenAPI for Subway ornage line (3호선)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "<your-url or ip>"
        }
    ],
    "paths": {
        "/station/eta": {
            "get": {
                "tags": [
                    "station"
                ],
                "summary": "3호선 역의 도착 정보를 알려줍니다. 역 정보 조회시 다음 규칙을 따릅니다.\n- 역이름 뒤에 오는 '역' 글자는 제외합니다.\n- 오금역 조회시, /eta/오금 을 사용합니다.\n- 고속터미널역 조회시, /eta/고속터미널 을 사용합니다.",
                "description": "3호선 지하철(Orange line) 역 도착 정보를 알려줍니다.",
                "operationId": "getOrangeLineInfo",
                "parameters": [
                    {
                        "name": "station_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Station Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/": {
            "get": {
                "summary": "Root",
                "operationId": "root__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}